Allow local plugin tools through decorators
