+++ specviz/utils/__init__.py
@@ -2,4 +2,5 @@ import os
 
 
 UI_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "data", "ui"))
+DATA_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "data"))
